// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Parse string date to type Date */
  CacheDate: any;
  /** The javascript `Date` as string. Type represents date and time as the ISO Date string. */
  DateTime: any;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: any;
};
// ...
export type AddressType = {
  __typename?: 'AddressType';
  id: Scalars['ID'];
  name: Scalars['String'];
  zip_code: Scalars['String'];
  street_address: Scalars['String'];
  street_number: Scalars['Float'];
  neighborhood: Scalars['String'];
  complement?: Maybe<Scalars['String']>;
  reference_point?: Maybe<Scalars['String']>;
  city_id: Scalars['String'];
  active: Scalars['Boolean'];
  created_at: Scalars['CacheDate'];
  updated_at: Scalars['CacheDate'];
  city: City;
};
// ...
export type Company = {
  __typename?: 'Company';
  id: Scalars['ID'];
  business_name: Scalars['String'];
  fantasy_name: Scalars['String'];
  email: Scalars['String'];
  phone_number: Scalars['String'];
  site: Scalars['String'];
  register_number: Scalars['String'];
  register_number_region: Scalars['Float'];
  address_one: Scalars['String'];
  address_two: Scalars['String'];
  zip_code: Scalars['String'];
  count_challenges: Scalars['Float'];
  challenges_subscription_count: Scalars['Float'];
  challenges_award_count: Scalars['Float'];
  company_type_id: Scalars['String'];
  city_id: Scalars['String'];
  active: Scalars['Boolean'];
  profile: Profile;
  company_addresses?: Maybe<Array<CompanyAddressType>>;
  created_at: Scalars['DateTime'];
  updated_at: Scalars['DateTime'];
};

export type CompanyAddressType = {
  __typename?: 'CompanyAddressType';
  address_id: Scalars['String'];
  company_id: Scalars['String'];
  is_billing_address: Scalars['Boolean'];
  address: AddressType;
  company: Company;
};
// ...
export type GetProfileDetailInput = {
  profile_id_requesting: Scalars['String'];
  profile_id_accessed: Scalars['String'];
};
// ...
export type Mutation = {
// ...
  updateProfileCompany: Profile;
// ...
};
// ...
export type MutationUpdateProfileCompanyArgs = {
  data: UpdateProfileCompanyInput;
};
// ...
export type Profile = {
  __typename?: 'Profile';
  id: Scalars['ID'];
  username: Scalars['String'];
  followers_count: Scalars['Float'];
  following_count: Scalars['Float'];
  official: Scalars['Boolean'];
  profile_avatar?: Maybe<Scalars['String']>;
  profile_cover?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  user_id: Scalars['String'];
  company_id: Scalars['String'];
  is_follower?: Maybe<Scalars['Boolean']>;
  user: User;
};

export type ProfileSchedule = {
  __typename?: 'ProfileSchedule';
  _id: Scalars['ID'];
  profile_id: Scalars['String'];
  resource_key: Scalars['String'];
  date_occurrence?: Maybe<Scalars['DateTime']>;
  title: Scalars['String'];
  subtitle: Scalars['String'];
  deep_link: Scalars['String'];
  web_link: Scalars['String'];
  image_link: Scalars['String'];
};

export type Query = {
// ...
  getProfile: Profile;
// ...
};
// ...
export type QueryGetProfileArgs = {
  data: GetProfileDetailInput;
};
// ...
export type UpdateProfileCompanyInput = {
  description?: Maybe<Scalars['String']>;
  profile_avatar: Scalars['String'];
  profile_cover: Scalars['String'];
  username: Scalars['String'];
  address_one: Scalars['String'];
  address_two: Scalars['String'];
  zip_code?: Maybe<Scalars['String']>;
  city_id?: Maybe<Scalars['String']>;
  fantasy_name: Scalars['String'];
  business_name: Scalars['String'];
  register_number: Scalars['String'];
  site: Scalars['String'];
  company_id: Scalars['String'];
  profile_id: Scalars['String'];
  phone_number: Scalars['String'];
};
// ...
export type User = {
  __typename?: 'User';
  id: Scalars['ID'];
  name?: Maybe<Scalars['String']>;
  firstname?: Maybe<Scalars['String']>;
  lastname?: Maybe<Scalars['String']>;
  email?: Maybe<Scalars['String']>;
  gender?: Maybe<Scalars['String']>;
  stature?: Maybe<Scalars['Float']>;
  weight?: Maybe<Scalars['Float']>;
  phone?: Maybe<Scalars['String']>;
  date_of_birth?: Maybe<Scalars['CacheDate']>;
  has_social_login: Scalars['Boolean'];
  city_id?: Maybe<Scalars['String']>;
  city?: Maybe<City>;
  active: Scalars['Boolean'];
  address_line_one?: Maybe<Scalars['String']>;
  address_line_two?: Maybe<Scalars['String']>;
  street_number?: Maybe<Scalars['String']>;
  legal_registry_number?: Maybe<Scalars['String']>;
  zip_code?: Maybe<Scalars['String']>;
  strava_permission_activities: Scalars['Boolean'];
  staff: Scalars['Boolean'];
  blacklist: Scalars['Boolean'];
  team_name?: Maybe<Scalars['String']>;
  companies: Array<UsersCompanies>;
  payments: Array<Payment>;
  created_at: Scalars['DateTime'];
  updated_at: Scalars['DateTime'];
  profile?: Maybe<Profile>;
  last_challenge?: Maybe<Challenge>;
  count_challenges_participates?: Maybe<Scalars['Float']>;
  user_event_extraordinary_action?: Maybe<Array<UserEventExtraordinaryActionType>>;
  activities_count?: Maybe<Scalars['Float']>;
};